requirements.txt
Flask==1.1.2
gunicorn==20.0.4
grpcio==1.26.0rc1
tensorflow==2.0.0b1
google-cloud-storage==1.2.0
scikit-learn==0.19.0
numpy==1.16.1
scipy==0.19.1
Pillow==2.2.1